+++ spec/features/staff/users_can_view_reports_spec.rb
@@ -59,7 +59,33 @@ RSpec.feature "Users can view reports" do
         click_on t("default.link.show")
       end
 
+      expect(page).to have_content report.financial_quarter
+      expect(page).to have_content report.fund.source_fund.name
+    end
+
+    scenario "the report includes a summary" do
+      report = create(:report, :active, organisation: build(:delivery_partner_organisation))
+
+      visit reports_path
+
+      within "##{report.id}" do
+        click_on t("default.link.show")
+      end
+
       expect(page).to have_content report.description
+      expect(page).to have_content l(report.deadline)
+      expect(page).not_to have_content t("page_content.report.summary.editable.#{report.editable?}")
+      expect(page).to have_content report.organisation.name
+    end
+
+    context "when there is no report descripiton" do
+      scenario "the summary does not include the empty value" do
+        report = create(:report, :active, organisation: build(:delivery_partner_organisation), description: nil)
+
+        visit report_path(report.id)
+
+        expect(page).not_to have_content t("form.label.report.description")
+      end
     end
 
     scenario "they can view budgets in a report" do
@@ -250,14 +276,12 @@ RSpec.feature "Users can view reports" do
 
     context "when there is an active report" do
       scenario "they can view reports for their own organisation" do
-        report = create(:report, :active, organisation: delivery_partner_user.organisation)
-        other_organisation_report = create(:report, :active)
+        _report = create(:report, :active, organisation: delivery_partner_user.organisation)
+        _other_organisation_report = create(:report, :active)
 
         visit reports_path
 
         expect(page).to have_content t("page_title.report.index")
-        expect(page).to have_content report.description
-        expect(page).not_to have_content other_organisation_report.description
       end
 
       scenario "can view their own report" do
@@ -269,7 +293,23 @@ RSpec.feature "Users can view reports" do
           click_on t("default.link.show")
         end
 
+        expect(page).to have_content report.financial_quarter
+        expect(page).to have_content report.fund.source_fund.name
+      end
+
+      scenario "their own report includes a summary" do
+        report = create(:report, :active, organisation: delivery_partner_user.organisation)
+
+        visit reports_path
+
+        within "##{report.id}" do
+          click_on t("default.link.show")
+        end
+
         expect(page).to have_content report.description
+        expect(page).to have_content l(report.deadline)
+        expect(page).to have_content t("page_content.report.summary.editable.#{report.editable?}")
+        expect(page).not_to have_content report.organisation.name
       end
 
       scenario "the report shows the total forecasted and actual spend and the variance" do