+++ CHANGELOG.md
@@ -662,6 +662,10 @@
 - Add "sdgs_apply" to CSV export
 - Allow organisations of type "External Income Provider" to be created
 - Move Aid type question before Collaboration type question
+- Report title no longer includes the description
+- A report description is no longer required
+- Reports show a more detailed summary, including deadline, organisation and
+  description (if one is supplied)
 
 ## [unreleased]